Auto csv import (#3)
* fix formats

* fix formats

* untrack ignored files for real

untrack ignored files

csv importer

csv importer

remove environment from docker-compose.yml

csv importer

* adjust db connection & file path for docker

* clean-up. csv importing works

* fixed id from string to int

+++ pokemon-app/src/main/java/com/example/springbootdocker2/repository/DeckRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface DeckRepository extends JpaRepository<PokemonCardDeck,Long> {
+public interface DeckRepository extends JpaRepository<PokemonCardDeck,Integer> {
 }
